authentication/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.4
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
pycparser==2.21
python-decouple==3.8
Werkzeug==2.3.7
